% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_odmap.R
\name{create_odmap}
\alias{create_odmap}
\title{Create folder with ODMAP structure}
\usage{
create_odmap(path = ".", name, overwrite = FALSE)
}
\arguments{
\item{path}{\verb{[character]} \cr Path where the folder will be created, including
intended folder name. If no argument is provided, the structure is created in
the current working directory,}

\item{overwrite}{\verb{[logical(1)=FALSE]} \cr Should the folder be overwritten if'
it already exists? Default is \code{FALSE}.}
}
\description{
This function creates an empty folder following the ODMAP structure
(Overview, Data, Modeling, Assessment, and Prediction). It is intended to
be used for habitat, movement, and connectivity modeling and suggested to
be used by species (i.e. one folder per species), but this can be used
in other ways depending of the user preferences.
}
\examples{
create_odmap(path = "./test")

}
