% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_rsf_habitat.R
\name{data_prep_rsf_habitat_rein}
\alias{data_prep_rsf_habitat_rein}
\title{Prepare data for RSF analyses for wild and semi-domesticated reindeer within SAM}
\usage{
data_prep_rsf_habitat_rein(
  dat,
  season,
  prediction = FALSE,
  fixwind = TRUE,
  land_cover_factor = FALSE,
  land_cover = c("norut_smd", "norut", "smd", "nmd")[1],
  ref_landcover = "heathland",
  include_zoi_nearest = FALSE,
  species = c("wrein", "trein")[1],
  prefix = "",
  reference_year = lubridate::year(lubridate::now()),
  use_binary_vars_as_categorical = FALSE
)
}
\arguments{
\item{dat}{\verb{[data.frame]} \cr Data set annotated with environmental covariates, almost
ready for analysis.}

\item{season}{\verb{[string]\{"sum", "cal", "win"\}} \cr Season of interest.
One of "sum", "cal", or "win".}

\item{prediction}{\verb{[logical(1)=FALSE]} \cr Additional changes that should
only be done in the grid for prediction.}

\item{land_cover_factor}{\verb{[logical(1)=FALSE]} \cr Logical variable stating whether
the land cover variable should be treated as a factor (one single column with
land cover type names) or not (as multiple columns as dummy variables).}

\item{land_cover}{\verb{[string(1)="norut_smd"]\{"norut_smd", "norut", "smd", "nmd"\}} \cr
Which land cover map should be used for analysis. It prepares the corresponding
classes, using the \code{ref_landcover} as the reference class.}

\item{ref_landcover}{\verb{[string="heathland"]} \cr Reference class for land cover
variables.}

\item{include_zoi_nearest}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, variables representing
the zone of influence (ZOI) of the nearest feature are also computed, based on
the distance to the nearest features. Only relevant if \code{prediction = TRUE}.}

\item{species}{\verb{[string="wrein"]\{"wrein", "trein"\}} \cr The species/management system
for which we should prepare the data for. Either \code{"wrein"} for wild reindeer or
\code{"trein"} for semi-domesticated reindeer. It only implies some differences in data
which are particular to each of the species/management systems.}

\item{prefix}{\verb{[string=""]} \cr Prefix to be added to the variable/column names
of the dataset. Default is \code{""}, but it could be e.g. \code{"startpt_"}, \code{"endpt_"},
or \code{"along_"} for variables extracted at the starting or ending point or along
steps.}

\item{use_binary_vars_as_categorical}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, binary
variables (e.g. lakes, reservoirs) should be treated as categorical. Otherwise,
they are treated as numerical.}
}
\description{
This function takes in the already annotated data for RSF with either wild or semi-domesticated
reindeer and further prepare columns and variables for the model fitting procedure.
The function is suited for point resource selection functions (RSF) only.
}
