% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_ssf_movement.R
\name{data_prep_ssf_movement_rein}
\alias{data_prep_ssf_movement_rein}
\title{Prepare data for SSF movement/permeability analyses for wild and semi-domesticated reindeer within SAM}
\usage{
data_prep_ssf_movement_rein(
  dat,
  season,
  prediction = FALSE,
  fixwind = TRUE,
  land_cover_factor = FALSE,
  land_cover = c("norut_smd", "norut", "smd", "nmd")[1],
  ref_landcover = "heathland",
  include_zoi_nearest = FALSE,
  species = c("wrein", "trein")[1],
  prefix = c("along_", "endpt_", "")[1],
  prefix_cross = "cross",
  reference_year = lubridate::year(lubridate::now()),
  use_binary_vars_as_categorical = FALSE,
  formula = NULL
)
}
\arguments{
\item{dat}{\verb{[data.frame]} \cr Annotated data set for analysis.}

\item{season}{\verb{[string]\{"sum", "cal", "win"\}} \cr Season of interest.
One of "sum", "cal", or "win".}

\item{prediction}{\verb{[logical(1)=FALSE]} \cr Additional changes that should
only be done in the grid for prediction.}

\item{land_cover}{\verb{[string(1)="norut_smd"]\{"norut_smd", "norut", "smd", "nmd"\}} \cr
Which land cover map should be used for analysis. It prepares the corresponding
classes as}

\item{include_zoi_nearest}{\verb{[logical(1)=FALSE]} \cr By default, \code{FALSE}. It should be
set to \code{TRUE} if there are variables representing the ZOI of the nearest feature in
the formula. Relevant only for prediction (using the grid).}

\item{prefix}{\verb{[string(1)="endpt_]\{"endpt_", "along_", ""\}} \cr Prefix for the variable,
whether \code{"endpt_"} or \code{"along_"} for variables measures in the end point or along
the step, and \code{""} for normal RSF or for the prediction using the grid.}
}
\description{
This function takes in the already annotated data for SSF with either wild or semi-domesticated
reindeer and further prepare columns and variables for the model fitting procedure.
The function is suited for step resource selection functions (SSF) only.
}
