% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_make_query}
\alias{db_make_query}
\title{Helper function to prepare queries to be sent to PostgreSQL}
\usage{
db_make_query(...)
}
\arguments{
\item{...}{\verb{[character]} \cr Strings representing the parts of a query.}
}
\value{
A string with a query, to be used, for instance, within \code{DBI::dbExecute()}.
}
\description{
Helper function to easily concatenate strings and prepare queries for
PostgreSQL.
}
\examples{
conditions <- paste0("\"KKOD\" IN ('5011', '5012', '5016', '5017')")

query1 <- db_make_query(
"CREATE TABLE IF NOT EXISTS sam_env.roads_public_se AS
SELECT *
FROM
\"Topography\".\"Sweden_Vagkartan_Roads_lines\"
WHERE ", paste(conditions, collapse = "AND "), ";")
query1
# to execute, first one needs a connection 'con' to PostgreSQL, then it can be run as
# DBI::dbExecute(con, query1)

}
