% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_write_vect}
\alias{db_write_vect}
\title{Helper function to upload vector layers to PostGIS database}
\usage{
db_write_vect(vect, con, dsn, pkey_column = "pkey", comment = "")
}
\arguments{
\item{vect}{Vector object in R, in \code{sf} format.}

\item{con}{Connection to PostGIS.}

\item{dsn}{\verb{[character(1)]} \cr Name of the layer after saved in the PostGIS
database. It might contain the name of the schema and the table, in the format
\code{"schema_name.table_name"}.}

\item{pkey_column}{\verb{[character(1)="pkey"]} \cr Name of the primary key column
to be created. Default is \code{"pkey"}.}

\item{comment}{\verb{[character=""]} \cr Comment to be added to the layer. Typically a
description of the layer, how it was processed, the source, and other relevant
information.}
}
\value{
The function does not return anything, but writes de vector into the
PostGIS database.
}
\description{
Helper function to easily upload vector layers to PostGIS, while adding a primary/unique
key and possibly a comment to the layer. This function also ensures that the owner
of the layer has automatically write to modify the layer as wanted.
}
