% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_along.R
\name{extract_along}
\alias{extract_along}
\title{Extract values of raster using in parallel}
\usage{
extract_along(
  r,
  sl,
  ws,
  col_step_length = "dist",
  step_id = "use_ava_data_animals_id",
  prefix = "along_"
)
}
\arguments{
\item{r}{\verb{[SpatRast]} \cr Set of rasters, in the \code{SpatRaster} format from \link{terra}.}

\item{sl}{Vector of step lines, with a column representing the step id
(stratum or just step number) and a LINESTRING geometry for each step.
Can be a \code{sf} or a \code{SpatVector} object.}

\item{ws}{List of which summary statistics are to be computed. Currently, only "mean", "max",
"min", and "sum" are implemented.}

\item{col_step_length}{\verb{[character="dist"]} \cr String with the name of the
column representing step length.}

\item{step_id}{\verb{[character="use_ava_data_animals_id"]} \cr String with the name
of the column representing step ID.}

\item{prefix}{\verb{[character="along_"]} \cr Prefix to be added to the extracted
variable names.}
}
\description{
Function to extract values of rasters along lines using parallel computation.
The function is actually not very optimized currently.
}
