% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_get_res.R
\name{grass_get_res}
\alias{grass_get_res}
\title{Get resolution from a raster map in a GRASS GIS database}
\usage{
grass_get_res(map)
}
\arguments{
\item{map}{\verb{[character]} \cr Name of the raster layer, possibly including the
mapset in the format \code{"map_name@mapset_name"} if in a mapset different from the
current mapset.}
}
\value{
A number representing the pixel size, based on the \code{nsres} parameter from
\code{r.info}, in units defined based on the project/location CRS. Should generally be meters.
}
\description{
Get resolution from a raster map in a GRASS GIS database
}
\examples{
# connect to GRASS
NinaR::grassConnect()
# get resolution
grass_get_res(map = "master_grid_100m_norway@p_sam_tools")
grass_get_res(map = "master_grid_1km_norway@p_sam_tools")

}
