% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_set_region.R
\name{grass_set_region}
\alias{grass_set_region}
\title{Set the region in a GRASS GIS session}
\usage{
grass_set_region(flags = c("p"), ...)
}
\description{
To know more about the arguments, check: https://grass.osgeo.org/grass82/manuals/g.region.html
}
