% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_raster_info.R
\name{grass_raster_info}
\alias{grass_raster_info}
\title{Get information from a raster}
\usage{
grass_raster_info(map)
}
\arguments{
\item{map}{\verb{[character]} \cr Name of the raster layer, possibly including the
mapset in the format \code{"map_name@mapset_name"} if in a mapset different from the
current mapset.}
}
\value{
A list with 9 elements, extracted from \code{r.info}:
the four coorinates defining the bounding box
of the raster (n, s, e, w); the raster resolution, taken as nsres; the number
of rows, cols, and cells, and the type of data (e.g. CELL, FCELL).
}
\description{
This function gets information from a raster layer in GRASS GIS and returns
it as a list. It is a wrapper for \code{\link[rgrass:r.info]{rgrass::r.info()}}.
}
\examples{
if(FALSE) {
  # connect to GRASS
  NinaR::grassConnect()
  # get raster parameters
  grass_raster_info(map = "master_grid_100m_norway@p_sam_tools")
}

}
